Refactor: separate AsyncRequests from Scanner
